Consolidate comprehensive test improvements into existing files
- Merged key improvements from comprehensive test files into existing register.spec.ts and profile-edit.spec.ts
- Removed duplicate comprehensive test files that were conflicting with main branch
- Improved OAuth testing with route interception and better error handling
- Updated login helper to use proper testid (quick-login-volunteer-button)
- Reduced timeouts for faster test execution (500ms → 300ms)
- Added test skipping for OAuth providers not configured in test environment

This consolidates all the test improvements while avoiding duplication and conflicts.
